Removed score from PfsParameter (brain dead!)

+++ jpa-model/src/main/java/com/wci/umls/server/jpa/helpers/SearchResultJpa.java
@@ -33,7 +33,7 @@ public class SearchResultJpa implements SearchResult {
   private boolean obsolete;
 
   /** The score. */
-  private float score;
+  private Long score;
 
   /**
    * Default constructor.
